added errorhandling and modularity for rps minigame
